cost fix

+++ sokoban-solver/src/back/interfaces/CostFunction.java
@@ -1,5 +1,5 @@
 package back.interfaces;
 
 public interface CostFunction {
-    public int evaluate(Game current, Game child);
+    public float evaluate(Game current, Game child);
 }
